[breaking] Always return React.createElement nodes on server and client
This is required if you want to compose multiple react components together on the
server. The template that you actually write out to the HTTP response should only
be rendered as a HTML string. This is now done by using a `{ html: true }` option
with the retruned template render function.